hlsltool: Add check that w != 1.0 in --auto-fix-vertex-halo
The real secret of the auto halo fix is that we need to match the driver
heuristics - if the driver adjusts the output position, we need to
adjust anything else based on it.

One of the heuristics that the driver does use is a W == 1 test to avoid
adjusting certain full screen effects. It's not known if that works in
isolation or only in conjunction with other conditions (like depth
buffer tests), but let's just try it out for a while and see what
happens.