MGSVTPP: Run --auto-fix-vertex-halo
Given the success of this in DX9 games, let's try it here and see what
happens. At first glance there is nothing new broken (only thanks to the
the W != 1 test though). Not much new fixed either, but given I already
manually fixed the halos in this part of the game that is to be
expected.