MGSVTPP: Fix another light pattern - masked lights
This pattern is fairly simple and all adjustments go in vertex shaders,
but can be a little difficult to find the right shaders.

1. Find the light shader that turns the scene blue when skipped and has
the depth buffer. Do not modify this shader.

2. Check the frame analysis logs looking for that shader and look for a
2nd shader pair that is interleaved (alternatively I think this is
another pair that turns it blue when hunting). This is the mask shader -
apply a standard stereo correction to the VS

3. For the VS paired with 1, adjust only TEXCOORD1 as part of the halo
fix, either by relocating TEXCOORD0 write, or using a 1/2 output
adjustment to TEXCOORD1.

To check if the shadow mask is right, set o0.w=0 in the PS found in 1,
which will show where the mask lies - you want this in 3D.