MGSV: Remove shader damaged by the decompiler
Noticed a subtle difference of this one in 2D with show_original -
slight difference in the shadow on the Doctor's coat, nothing major.
This had an auto halo fix applied but it doesn't appear to have made a
visible difference.

Kill it for now, no point including it if it's damaged and doesn't
improve anything.