shadertool: Add support to use the undocumented stereo register
nvidia injects the stereo parameters in c255 (by default, wouldn't
surprise me if a profile setting can change that), and we can use that
to our advantage to apply a stereo correction without needing to use on
of the coveted sampler registers.

The register's contents is not quite the same as the stereo sampler, so
we need to adjust the maths a little (worked out with a little trial and
error in the DX11 equivalent in vs-cb12):
x = -separation * convergence
y = separation

That means, with a patched version of Helix Mod we should be able to
avoid breaking Ceto water and still be able to fix the effects.