The Forest: Use undocumented register in manually fixed vertex shaders
This starts removing the dependency on the Helix Mod stereo sampler from
the vertex shaders as a step towards the goal of fixing the conflict
between Helix Mod and Ceto water. It replaces the Helix stereo sampler
with the nvidia stereo register and adjusts the maths to work with the
available values.