The Forest: Revert light shaft shaders where the stereo register does not work
It seems that not all vertex shaders have access to the undocumented
stereo register, which unfortunately is a bit of a deal breaker for the
goal of applying a fix without changing any of the vertex shader sampler
registers.

It might still be possible if we can replicate the stereo nature of the
constant registers with another tool using nvapi calls, but we cannot do
this with Helix Mod and it's a little too involved to binary patch in.

This knowledge may still be useful for other purposes however, like
determining if the driver heuristics were going to apply a stereo
correction (I can think of a few shaders where this could be helpful),
provided that we can rely on the driver always using c255 (which given
that it's undocumented is not a given).