The Forest: Redo all autofixed vertex shaders to use the nv stereo register
This is a bit of a dead end for now since we can't use the stereo
register in the light shaft & sky glow shaders, but we may be able to
overcome that later. For now change the remaining vertex shaders to use
the stereo register to prove out that it is safe for vertex halo and
reflection fixes.