Shadertool: Update help documentation on the limitations of using the nv register
Experiments have shows that the undocumented register containing the
stereo settings from the nvidia driver is only set when the driver's
heuristics decide to adjust a vertex shader.

That means that the register cannot be relied upon, so I do not
recommend using it (besides which, given that it is not documented we
have no way to know if it could change in a new driver version, profile
update, etc).

Having said that, I still have some ideas where this might come in handy
- it could be a useful indicator that the driver is applying a stereo
correction for shaders where it is applied inconsistently, and we might
still be able to use something similar with a new tool that calls nvapi
to set our own stereo register since it is clearly supported.

But at least for my hopes of using this with a binary patched Helix Mod
to fix Ceto water this is a dead end.