The Forest: Update README
Ceto water is now partially broken in this game as well - the waves
activate inconsistently and there seems to be some slight differences
between the eyes.

I went down a bit of a rabbit hole to see if I could use the driver's
stereo register instead of Helix mod's stereo sampler to fix it (as this
would allow a binary patched Helix mod to be used that does not set a
stereo sampler on the vertex shader), but that was largely a dead end as
the stereo register is not available in all shaders (in particular, it
would have re-broken light shafts).