MGSGZ: Fix depth of flare around lights
The lens flare shader does have the correct depth available, but it is
not used in the position calculation (it is only passed through to the
pixel shader for comparison with the depth buffer for opacity
calculations). That means we can adjust the flares to the correct depth
relatively easily.