MGSGZ: Add an auto-crosshair
Add my crosshair.hlsl from FC4 and use it to automatically adjust the
crosshair to the correct depth. Obtaining the depth buffer is easy as
the game seems to leave it assigned while drawing the HUD.

stereo2mono is not working for some reason that I'll need to
investigate, but the accuracy improvement it brings probably won't add
much as this is a 3rd person game - so long as the target doesn't get
obscured in one eye very often it should be fine.

The projection parameters are copied from the pixel shader and used to
scale the depth buffer.