MGSV: Add convergence override on aim + F combo
The F button is multi-purpose in this game, and needs different
convergence presets for each case.

While holding F without holding aim, increase convergence for the
binoculars

When tapping F while holding aim, decrease convergence for the first
person view

To achieve this we use two ini params as condition variables - x is
already set while aiming, so we can use that, and set w to the opposite
of x.

We also need to add a convergence preset to the aiming button so that
the original convergence will be restored when it is released.