MGSVTPP: Fix fake mirrors rendering in one eye
Using stereo2mono to invert the mirrors doesn't work for some of them
that use a static texture instead of a render target. Instead, this
results in the fake reflection only rendering in one eye.

The fake reflections all seem to be square, so check the Texture2D
dimensions in the shader to decide whether to use the stereo2mono copy
or not. This works for the first few mirrors at least, but needs to be
confirmed if there might still be any problematic mirrors elsewhere (if
there are, it would be fairly straight forward to add a feature to put
the format in an ini param and check that instead, or possibly by
checking if initialdata was used when the resource was created).