MGSTPP: Fix more specular highlights
Fixes some on the main menu and another light in the hospital. The
pattern only seems to apply to some of the lighting shaders, not all -
look for a pattern similar to this just below the shadow adjustment:

r0.x = dot(view-space-coord.xyz, view-space-coord.xyz);
r0.x = rsqrt(r0.x);
 = view-space-coord.xyz * r0.xxx;

The use of the view space coordinate is key here, avoid if you see that
pattern applied after any mention of g_psLight.m_lightParams.