MGSTPP: Add texture filtering for crosshair & floating icons, add aiming overrides
Same type of aiming convergence overrides as GZ, including the
conditional to get 1st person convergence working.

Add texture filtering + aiming conditional for the four dot crosshair
adjusting based on the center of the screen.

Add texture filtering and non-aiming conditional for the pulsating
circle and square that appears over the target. Currently these are
adjusted based on their vertex position - I need to decode the vertex
buffer to try to adjust it based on the icon center.