MGSVTPP: HUD tweaks (on the back-burner due to texture filtering)
This is whatever I'm up to tweaking the HUD. I'm putting attempts to do
anything advanced with texture filtering on the back-burner for now -
this game creates the same type of hash contamination we saw in FC4,
making texture hashes unreliable without track_texture_updates.

track_texture_updates is rather expensive though, which is currently
unavoidable due to the way the texture hashes are designed. I have some
ideas to reduce this cost, but it will involve changing the texture
hashes, so I don't want to spend much time collecting hashes that are
likely to change in the future.

For now use the tweaks that don't depend on texture hashes, and only use
them for the crosshair (and it's not clear that even that is reliable)
in combination with the aim button. Forget about adjusting floating
icons for now.