MGSTPP: Add toggle to kill broken blue enemy highlight
The blue highlights of enemies behind cover is broken around the hips. I
spent some time looking at this, but it seems to be a game bug (broken
bone matrices I think) that is obvious in 3D, but not as apparent in 2D.
Kill the effect and add a toggle on F3 to bring it back (it's not
unplayable, just not great).