MGSTPP: Fix opacity of waypoint circle on map
The waypoint circle opacity varies with distance from the center of the
screen, but it needs to be the center of the map. Pass the offset that
the driver + out shader will have adjusted the circle through to the
pixel shader and use that to correct the position used for opacity.