MGSTPP: Fix disappearing menu
After progressing a little way into the game, the main menu changes to
use a 3D menu, but it's a lot closer than the other 3D HUD elements.
Therefore, add a 2nd custom convergence for it and a cut-off depth
between the two.

The cut-off is currently set to 48 so that the DD symbol on the mission
select screen uses the low convergence, and hopefully will allow the
name + birthday screen to use the high convergence.

This also reassigns y to y1 so that all four HUD depth related
parameters can be loaded from a single vector in IniParams.

Reported-by: mrorange