MGSTPP: Fix flickering near screen edge of sand decals
The blowing sand decals were flickering at the edges of the screen. It
seems they are a post processed effect and should cover the entire
screen, but the driver's heuristics couldn't make up their mind.

Normalise the output from the vertex shader such that W == convergence
to neutralise the driver heuristics and remove the flicker.