MGSTPP: Remove clipping on small band on Horizon above Ocean
In GZ the entire ocean had clipping, but here only a small band on the
horizon does. This is actually because there is a second layer that is
not clipped in front of it, and removing that shows the same clipping we
had in GZ, and the same pattern fixes it here as well.