Move 3D HUD cut-off back to unbreak name+birthday screen
I had the cutoff where it was for the DD symbol, but the name+birthday
screen was broken as a result. I can't move it forward without breaking
the DD symbol again, so move it back until the name+birthday screen is
entirely using the low convergence. Not ideal, but a better compromise
for now. If it gets more complicated I'll probably have to use scene
detection or give up on it.