MGSTPP: Fix another lighting shader
Tip: This shader was not obvious when skipped - it removed shadows, but
left the specular highlights. The obvious shader that removed both was
too simple, but checking the frame analysis log showed this as the very
next shader used after that one. Might be a handy way to identify
difficult to hunt lighting shaders.