MGSTPP: Auto adjust red dot sight like crosshair
The red dot sight is lined up with the weapon scope, but it's nicer to
line it up with the target like a crosshair.

We could reuse the same intermediate resource as the HUD shader, but the
log shows it being constantly destroyed and recreated if we do. That
will work, but is not ideal, so use a second intermediate resource.