Demonicon: Fix shadows
This game needs a two stage shadow fix, similar to the new Unity
world-space fix. It uses both the MVP and the M2WS matrices and their
inverses to convert between world space and projection space and back
again.

Also similar to Unity, a correction must be applied in the VS as well
(uncorrecting the VS instead gives a quick starting point, but causes
clipping - to solve both, the VS must be corrected instead).

Specular highlights are floating a little above the surface - haven't
tracked these down yet.