MGSTPP: Fix misalignment on sky near horizon when camera is tilted
When the chopper takes off from Mother base the sky can be seen to be
misaligned out near the horizon. I wasn't sure if it was just crosstalk
at first, but it turns out that it was really misaligned and this
adjustment to the sky shader seems to do the trick.