The Witness: Auto adjust the mouse cursor depth
Not everything that can be interacted with is at the same depth, so
bring in the auto crosshair code to adjust the depth automatically.

Similar to FC4, we are able to find the center of the mouse by looking
up the position of two diagonally opposite corners in the vertex buffer.