The Witness: Substantial improvements to auto cursor for some puzzles
Some of the puzzles are drawn with the same shader that is responsible
for the white gate, etc. This shader is used with the depth buffer set
to read only, so the puzzle does not appear on the depth buffer and
therefore the cursor will not line up with it. This makes it rather
difficult to solve the puzzle with 3D enabled.

This commit solves this problem through numerous brand new features
introduced in 3DMigoto 1.2.28:

- This creates a new resource, which is made compatible with the depth
  target via the new copy_desc keyword. The format is overridden as the
  depth target's existing format cannot be used in a render target, nor
  cast to anything that could be.

- A custom shader to draw the depth information onto the new resource.
  This makes use of the new draw=from_caller feature to automatically
  use the same draw call as the game and combined with the fact that it
  does not override the vertex shader it will render the same geometry.
  The vertex shader is modified to pass the depth value through to the
  pixel shader.

- The custom resource is cleared on every present call by injecting
  another custom shader.

- The crosshair.hlsl shader has been modified to use the custom resource
  first, and fall back to the depth buffer for any blank pixels.

Additionally, the shader I was fetching the depth from is not drawn all
the time, so the auto cursor would not work when it was not active. I
now fetch the depth from the oD slot of another shader which is
responsible for scaling it (still need to confirm if that shader is used
on all quality settings and throughout the game).

Since the oD slot is using MSAA in this game on the default settings,
this forces the cursor vertex shaders to vs_5_0 using another new
feature of 3DMigoto to allow them to use Texture2DMS. For now this does
away with using stereo2mono as that does not yet work on MSAA resources.
The crosshair.hlsl shader has been modified to work with MSAA resources.