The Witness: Remove perspective divide from puzzle depth buffer
I had written to our puzzle depth buffer as z/w intending it to use the
same scaling as the real depth buffer, but this causes some artefacts at
certain depths, particularly if one of the vertices is off-screen or
behind camera.

I don't need to scale it, so just write the depth value to it directly
and have the crosshair.hlsl use that directly.