The Witness: Add key to kill separation for certain puzzles
It seems there are some puzzles in this game that rely on the 2D
perspective to show the solution, which does not work in 3D where it
will no longer line up. For these puzzles, add a key binding on Q to
temporarily reduce the separation to see the solution.