The Witness: Approximate fog fix
The fog jumped when getting close to the effect boundary, which was
caused by a misaligned world space output, however correcting that broke
the fog more, which needed an additional correction in the pixel shader
to realign.

This seems to only be approximate given that the fog does not look quite
as nice as it does from inside the effect (which uses a different set of
shaders and renders full screen). The transition between the two is also
noticeable.