The Witness: Remove old puzzle Z buffer
This copy was causing an extreme fps hit in certain areas, so remove it.
Should still be ok most of the time - since the shader is run without
writing to the depth buffer the game must already be either sorting the
draw order from back to front, never draws multiple effects using this,
or uses a blend mode that will work in either order.