The Witness: Improve reflections when camera is tilted (approximate)
This negates the world space Z adjustment to improve reflections when
the camera is tilted. This is only approximate - we really want to
mirror it with respect to camera up, but this still makes the end game
much easier to look at.

Just adjust the outdoor reflection shaders that will be shown tilted and
leave the underground ones alone to be safe.