hlsltool: Use fxc to validate that shaders will compile
Using hlsltool to offline fix a game carries a high risk that a broken
shader may be included, and in some cases may not even compile.

There is nothing automated we can do about shaders that are damaged but
still compile, but we can try compiling the shaders when they are
installed to catch any with obvious errors.

If the compilation fails, rename the shader to ~failed and add the
errors from the compiler in a comment at the bottom.