Firewatch: Change full screen detection to use any instead of all
Sometimes on the radio the game goes into letterbox mode, and the full
screen detection only works on the corners resulting in the letterbox
going into the scene. Change the full screen detection to check for a
match on any edge instead of only the corners to hopefully avoid
adjusting it at all.