Proof of concept to convert games to SBS / TAB
Here's a proof of concept converting a game to SBS / TAB with 3DMigoto.
This only solves one of the issues - the driver will still be driving
the display as 3D Vision (meaning the 720p limit still applies with 3D
TV Play) and if it cannot force SBS/TAB in that mode it will not work.
It also does not support full frame SBS/TAB, only half-packed. EDID
overrides may help, but I haven't tried. I am not working on this
actively, there is no ETA as to when or if it will be usable. Edit:
joker18 and DHR figured out an easy solution to the output problem by
using checkerboard mode, so this actually works now :)