3DVision2SBS: Skip pixel shader if mode = 0
If the mode is set to 0 we just pass the copy of the back buffer
straight through unmodified, so as a small optimisation, skip the pixel
shader in that case. A further optimisation would be possible when/if
conditional logic is added to the d3dx.ini to skip the back buffer copy.