3DVision2SBS: Disable rasterizer culling
Running a custom shader is subject to the state that the game has left,
which can vary between games, and even within a single game. One of the
problematic areas is what culling is enabled - front or back faces, and
which is defined by clockwise vertices.

Use the new rasterizer state override ability in 3DMigoto 1.2.32 to
disable culling to remove this variable and make it more reliable.