3DVision2SBS: Backup and clear all render + depth targets
If a game has a depth target assigned on the present call it might cause
us issues, and if it has any render targets that are not compatible
with the back buffer we can also run into problems. To be safe, back up
all 8 possible render targets and the depth target and clear all that we
aren't using.

This is excessive, but should be safe. Later we can implement a
shorthand for this in 3DMigoto to simplify it.