hlsltool: Re-enable optimisation in compilation validation step
It seems that there are certain errors that will only be produced by fxc
when optimisations are enabled. Since the only point of compiling the
shaders from this script is to find any that will fail to compile under
3DMigoto it is important that we get the same errors, so re-enable
optimisations.

Unfortunately this will make the script take longer to execute if
validation is enabled.