hlsltool: --auto-fix-vertex-halo: Restrict temporary register search components
If the output position does not have all it's components set, we would
still search for reads of the unrelated components of the temporary
register. While usually harmless, this could occasionally move the
correction into the middle of some flow control if one of the components
was used but was unrelated to the position.

Restrict the search for reads to the components that are copied to the
output position to minimise the chance of this occurring.