Far Cry Primal: Fix reflections via stereo2mono technique
The reflections in Far Cry Primal follow the same pattern as Far Cry 4,
but fixing those required changes to every single vertex shader that
might ever be used when creating the reflection, and there were a lot of
these and I still find more.

Instead, take a different approach of using 3DMigoto's stereo2mono
support to reverse the eyes of the reflection (similar to the mirror in
MGSVTPP), combined with the same world space correction and forced
surface creation mode result in accurate 3D reflections.

It seems that there are some issues with the reverse stereo blit and
resources with mip-maps - it seems that the reverse stereo blit only
takes effect on the most detailed mip map, and the rest just end up with
a copy in one eye. The reflection seems to combine several mip-maps
together, and if they are not present it will be much too dark.

To work around this, use a custom shader that reverses the reflection to
a custom resource with only a single mip-map level so that there will
only be one to sample. This makes the ripples more detailed than they
are supposed to be (looks fine to me), so I might revisit it later and
see what we can do to solve the reverse stereo blit more generally.