Introduce asmtool.py, a port of shadertool/hlsltool for DX11 assembly
asmtool aims to be the equivalent of shadertool & hlsltool, but for DX11
assembly shaders. Internally it works more like hlsltool than shadertool
to get the benefits of the cleaner approach I took with hlsltool. Like
hlsltool it imports various routines from both shadertool and hlsltool
for common code and as such cannot work alone.

It currently implements --auto-fix-vertex-halo, and in fact the new
design I went for in hlsltool/asmtool means that there is actually only
a single implementation (in hlsltool), with the Shader object
abstracting the differences between the two.

Shadertool still has a separate --auto-fix-vertex-halo implementation,
and while it would be possible to make it share the same implementation
it would take some refactoring work.