FCPrimal: Fix shadow volumes / light shafts (includes debugging shader)
The volumetric shadow volumes / light shafts needed a world-space
correction, but doing so distorted them crazily.

This commit includes a custom debugging shader that visually shows the
shadow volume slices and the 2D shadow maps they are derived from. This
made it much easier to see if each slice was being correctly adjusted,
and quickly showed that with the correction enabled the shafts were
being blurred, pointing the finger at the blending with the shadow
volume of the previous frame.

With that knowledge, use both corrected and uncorrected coordinates to
correct the light shafts and not mess up the blending.