FCPrimal: Fix Ambient Occlusion
This was a little tricky to find the right solution, but like other AO
shaders it is a matter of adjusting a number of coordinates throughout
the shader.

Nowadays with the ability to output intermediate results in the shader
live I have a little more insight into what is happening - the first
depth texture sampled is the depth of the current pixel, and the
subsequent depth samples have a bokeh blur for the AO.

Coordinates constructed from all need to be adjusted, but finding the
exact places to adjust them was a little tricky (especially as until
they are all correct the shader tends to break more). One trick I used
was to output the contents of the X coordinate directly and observe that
the correct adjustment pushed "0" (black) from screen depth to infinity
in each case.