FCPrimal: Fix velocity buffer output from hands (fixes AO smearing on hands)
The velocity buffer is populated during the forward rendering part of
the frame, but the velocity of certain objects is wrong - hands, goats,
etc.

This can be fixed by using the original 2D position, which then records
the correct velocity and the AO smearing is removed.

This is only one shader for now - this pattern will need to be applied
to many more to fix this completely.