FCPrimal: Disable velocity buffer use in AO
The velocity buffer is used to allow previous frame's AO to be used with
higher accuracy when the camera is moving, or when something is moving
on the frame.

However, the velocity buffer is pretty broken in the forward rendering
part of the frame, and while it is simple to fix as demonstrated by the
previous commit, a couple of hundred shaders would be needed to do so.
Later on I might script that in assembly, but for now just kill the
velocity buffer's use in AO as it does not seem to add much.