FCPrimal: Improve physical lighting fix (2 confirmed, 2 unconfirmed)
These were a little tricky as my first attempt proved to be inaccurate,
and other attempts also proved inaccurate.

Add the same depth buffer scaling as was used in the directional
lighting shader, confirmed to be correct by using a custom shader to
verify that when convergence = 1.0, anything deeper than screen depth
(depth >= 1.0) was displayed as white.

The stereo correction goes before the use of the inverse view-projection
matrix and includes the perspective divide.

Another debugging step was used to visualise the Z value after that
matrix, which shows a band around Z==0 which is pushed to infinity when
using the correct adjustment.

Renaming these from tile lighting to physical lighting - the headers
suggest it is tile lighting, but until I see the telltale clipping I'm
going with the term of what it actually seems to adjust.

This includes shaders for environment=normal & high, as well as another
two that seem to follow the same pattern, but are not confirmed.