FCPrimal: Disable texture filtering + more HUD tweaks
FCPrimal suffers the same issue as FC4 with texture filtering, and while
I can counter that with track_texture_updates, but given that I'm
planning to rework that let's hold off on adding texture filtering.

In the meantime, restrict the crosshair adjustment to the center of the
screen, and only from the shader responsible for the crosshair.

Add the depth buffer copy to other HUD shaders that were missing it &
fix the mip-map detection being inverted.