FCPrimal: Fix tile lighting clipping
This was indeed tile lighting, but the clipping artefacts were fairly
controlled outdoors and only became evident a little way into the game.

The pattern is fairly simple in this case - apply a world space
correction to LightCullingData.pos in one of the compute shaders that
feeds into the physical lighting compute shaders.