FCPrimal: Switch depth buffer copy to use ambient lighting shaders
Noticed that the depth buffer copy stopped working around sunset for
a short time until the moon rose. Switch to the ambient lighting shaders
to see if they fare any better. Note that the scaling used for these is
slightly different.

Also tried the physical lighting shaders, but the copy was not working,
which I'd need to investigate.

Alternatively I might use frame analysis and grab it from one of the
shaders that rescales it like I did in FC4.