FCPrimal: Remove conflicting assembly shader
Had both assembly & HLSL versions of this decal shader since the autofix
script had performed a halo fix on it, which is no longer required since
we adjust it earlier.

Added a quick check for these in the autofix script, but I probably need
something in mkrelease and/or the pre-commit hook and/or 3DMigoto itself.