FC4: Switch includes to be relative to the shader
The UPlay version of FC4 has the working directory in the bin directory,
while the Steam version is at the top level of the game. This makes
using an include relative to the current working directory impossible to
work on both.

Now that includes can work relative to the path of the shader, switch
all HUD shaders to use that as it should work on both versions.