FCPrimal: Add alternative auto HUD adjustment and toggle key
Using the crosshair depth for the entire HUD overall works pretty well -
to the point where I'm thinking I'll just leave it as is rather than go
and add the same type of filtering I used in FC4, especially given some
of the transparent menus the game uses.

However, the subtitles are problematic (particularly in cutscenes), so
add a second HUD mode that prefers to place the HUD in front of anything
else on the screen (excluding the edges so that the hands don't keep
making it jump while running) for use in those, and a toggle key to
switch between the two modes.