FCPrimal: Disable Z convergence adjustment
I found the high convergence with the owl a little uncomfortable, so
didn't end up using very much. Also, because of the way it works it is
likely to reset a custom convergence players might use, so just disable
it. The high convergence on the backslash key can still be used in the
Mammoth DLC (where it works well).