dx11shaderanalyse.py: Add --hash-chunks option to find identical shaders
MGSVTPP uses the same shaders in single player as online, but they were
compiled from a different location which causes their debug info to
differ, and therefore have different hashes.

This option can be used to hash individual sections of the shader so
they can be matched up based on only their bytecode.