dx11shaderanalyse: Add DXBC hash implementation
I've always been curious to see how they managed to stuff up MD5 -
whether it was just an honest mistake, or if it was intentional. To this
end I implemented the MD5 algorithm by following the pseudocode on
wikipedia translated to Python, then attempted to modify it in various
ways to try to get the same result as the d3dcompiler. This way, I could
understand *what* they changed in the algorithm more clearly than just
staring at optimised code from a disassembler/decompiler.

The fact that the hash is based on MD5 is blatantly obvious since it
uses the exact same constants (both the initial constants, and the sine
wave values), rotation values, etc. as MD5, but running MD5 over the
shader did not yield the same hash.

Upon closer examination, it turns out that the way the data is padded
with the message size is slightly different - the details are in a
comment in the code. In my opinion, this particular modification looks a
little too bizarre to have been an innocent mistake, leading to the
conclusion that they intentionally obfuscated the hash for some
misguided reason.