New script: rename2bytecode.sh
This script follows the same principle as rename2embedded.sh, and
renames the shaders to use a crc32c hash of their SHEX/SHDR section.

Since it is expected some originally identical shaders may be renamed to
the same hash, if there is a collision it will show the diff and ask
before overwriting.