unity_asset_extractor: Add support for extracting Unity 5.3 shaders
Unity 5.3 changed the way they store shaders - the .shader assets still
seem to be extracted ok, but they are missing the shaders themselves as
well as most of the metadata that we were after (notably they do still
seem to include indications of shadow casters).

The shaders are now actually located in an lz4 compressed blob after the
shader text, and this patch adds preliminary support to extract this. It
currently decompresses the blob and writes it to disk, but does not yet
pull the individual shaders out of the blob (which will need to either
be added to extract_unity_shaders.py, or a new script created).

Further analysis is required to decode the decompressed binary format as
that is new as well, and it contains the metadata that we need, which is
once again being stripped from the shaders.

Annoyingly, Unity neglected to bump their file version field or
otherwise add an indication that the format had changed, so this checks
the Unity version string in the file to determine if it was created with
Unity 5.3 or higher.