extract_unity53_shaders: Handle variable size header, Better DX9 support
The variable size header on the bind info is still a bit of a mystery,
but it seems to end in a reasonably consistent manner depending on the
DX version. Use that as the best guess to find the start of the bind
info clauses.

Some rarer types (VectorInt, ScalarBool) were also added - pretty sure
this can be more generic between the common ones and these, but for now
using asserts so I catch any variations of these.

With this change, this can now extract headers for all shaders in The
Long Dark, both DX11 and DX9. It still lacks a way to distinguish
between vertex and pixel shaders.