extract_unity53_shaders: Fix DX9 shader extraction
DX9 shaders don't have the 5 undeciphered bytes preceding the shader
that DX11 does, so separate out the two extraction paths.

With this change, DX9 shaders are extracted with the correct CRC. It
seems that Unity 5.3 is not stripping DX9 shaders, but is stripping DX11
shaders, so for DX9 we could potentially skip the bind info parsing, but
we will still need to add headers to denote shadow casters before we can
reliably script this.