extract_unity53_headers: Add headers from .shader file as well
We still need some of the info from the .shader file (most importantly
to determine which shaders are shadow casters), so parse it and add the
info at the top of the extracted headers.

As far as I can tell, the only correlation between the shaders in the
two files is their order.